Added tracker functionality to the 'move' method
In lieu of the architectural issues I have with this slider, I needed it for a quick project and the move method was not calling my tracker function.  I added the ability to have a tracker passed into the method and call it when animating so my slider values can update live.